changed some floats to double. worked on main program some more

+++ Generator.java
@@ -1,7 +1,7 @@
 
 public interface Generator {
 	public PowerType getPowerType();
-	public double getPowerGen(float t);
+	public double getPowerGen(double t);
 	public double getGenTime();
 	public boolean isWorking();
 }
